+++ jdaviz/core/user_api.py
@@ -224,8 +224,13 @@ class UserApiWrapper:
                 exp_obj.selected = value
                 return
             elif len(exp_obj.create_new.choices) > 0:
-                exp_obj.create_new.selected = exp_obj.create_new.choices[0]
-                exp_obj.new_label.value = value
+                if value in exp_obj.create_new.choices:
+                    # value matches a viewer type label (e.g. 'Table', 'Scatter') -
+                    # select that type and the label will also default to that type
+                    exp_obj.create_new.selected = value
+                else:
+                    raise ValueError(f"{value} is not a valid viewer type."
+                                     f"  Valid labels are: {exp_obj.create_new.choices}")
                 return
 
         if isinstance(exp_obj, SelectPluginComponent):